# base-backend/src/main/java/com/samkruglov/base/api/UserController.java
package com.samkruglov.base.api;

import com.samkruglov.base.api.config.Current;
import com.samkruglov.base.api.view.mapper.UserMapper;
import com.samkruglov.base.api.view.request.ChangeUserDto;
import com.samkruglov.base.api.view.request.CreateUserDto;
import com.samkruglov.base.api.view.response.GetUserDto;
import com.samkruglov.base.domain.User;
import com.samkruglov.base.service.UserService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;
import javax.validation.constraints.Email;

@RestController
@RequestMapping("/api/users")
@RequiredArgsConstructor
@Tag(name = "users")
@Validated
public class UserController {

    public static final String CREATE_USER_OP_ID = "create-user";

    private final UserService service;
    private final UserMapper mapper;

    @Operation(operationId = CREATE_USER_OP_ID)
    @PostMapping
    public void createUser(@Valid @RequestBody CreateUserDto userDto) {
        service.create(userDto);
    }

    @GetMapping("/self")
    public GetUserDto getMe(@Current User user) {
        return mapper.toGetUserDto(user);
    }

    @GetMapping("/{email}")
    public GetUserDto getUser(@Email @PathVariable String email) {
        return mapper.toGetUserDto(service.getByEmail(email));
    }

    @PutMapping("/self")
    public void changeMe(@Current User user, @Valid @RequestBody ChangeUserDto changeDto) {
        service.change(user, changeDto);
    }

    @PutMapping("/{email}")
    public void changeUser(@Email @PathVariable String email, @Valid @RequestBody ChangeUserDto changeDto) {
        service.changeByEmail(email, changeDto);
    }

    @DeleteMapping("/{email}")
    public void removeUser(@Email @PathVariable String email) {
        service.deleteByEmail(email);
    }

    @DeleteMapping("/self")
    public void removeMe(@Current User user) {
        service.delete(user);
    }
}